+++ src/eval/bc/stencils/call.c
@@ -19,9 +19,9 @@
 
 #include "abi.h"
 
-mino_stencil_chain_t stencil_op_call(mino_val_t **regs,
-                                      mino_val_t **consts,
-                                      mino_state_t *S)
+void stencil_op_call(mino_val_t **regs,
+                      mino_val_t **consts,
+                      mino_state_t *S)
 {
     regs = mino_jit_call_slow(S, regs,
                               (unsigned)IMM_A,