+++ src/eval/bc/stencils/closure.c
@@ -15,9 +15,9 @@
 
 #include "abi.h"
 
-mino_stencil_chain_t stencil_op_closure(mino_val_t **regs,
-                                         mino_val_t **consts,
-                                         mino_state_t *S)
+void stencil_op_closure(mino_val_t **regs,
+                         mino_val_t **consts,
+                         mino_state_t *S)
 {
     regs = mino_jit_closure_slow(S, regs,
                                  (unsigned)IMM_A,