+++ src/eval/bc/stencils/conj_vec.c
@@ -13,9 +13,9 @@
 #include "abi.h"
 #include "runtime_layout.h"
 
-mino_stencil_chain_t stencil_op_conj_vec(mino_val_t **regs,
-                                          mino_val_t **consts,
-                                          mino_state_t *S)
+void stencil_op_conj_vec(mino_val_t **regs,
+                          mino_val_t **consts,
+                          mino_state_t *S)
 {
     regs = mino_jit_conj_vec_slow(S, regs,
                                   (unsigned)IMM_A,