+++ src/eval/bc/stencils/env_bind.c
@@ -14,9 +14,9 @@
 
 #include "abi.h"
 
-mino_stencil_chain_t stencil_op_env_bind(mino_val_t **regs,
-                                          mino_val_t **consts,
-                                          mino_state_t *S)
+void stencil_op_env_bind(mino_val_t **regs,
+                          mino_val_t **consts,
+                          mino_state_t *S)
 {
     regs = mino_jit_env_bind_slow(S, regs,
                                   (unsigned)IMM_A,