+++ src/eval/bc/stencils/getglobal_cached.c
@@ -32,9 +32,9 @@
 #include "abi.h"
 #include "runtime_layout.h"
 
-mino_stencil_chain_t stencil_op_getglobal_cached(mino_val_t **regs,
-                                                  mino_val_t **consts,
-                                                  mino_state_t *S)
+void stencil_op_getglobal_cached(mino_val_t **regs,
+                                  mino_val_t **consts,
+                                  mino_state_t *S)
 {
     mino_bc_ic_slot_t *slot = &MINO_JIT_BC_IC_SLOTS(IMM_BC)[(unsigned)IMM_BX];
     mino_thread_ctx_t *ctx  = MINO_JIT_INVOKE_CTX(S);