+++ src/eval/bc/stencils/inc_i.c
@@ -13,8 +13,8 @@
 #include "abi.h"
 #include "runtime_layout.h"
 
-mino_stencil_chain_t stencil_op_inc_i(mino_val_t **regs, mino_val_t **consts,
-                                       mino_state_t *S)
+void stencil_op_inc_i(mino_val_t **regs, mino_val_t **consts,
+                       mino_state_t *S)
 {
     mino_val_t *v = regs[IMM_B];
     long long   r;