+++ src/eval/bc/stencils/le_ik.c
@@ -7,8 +7,8 @@
 #include "abi.h"
 #include "runtime_layout.h"
 
-mino_stencil_chain_t stencil_op_le_ik(mino_val_t **regs, mino_val_t **consts,
-                                       mino_state_t *S)
+void stencil_op_le_ik(mino_val_t **regs, mino_val_t **consts,
+                       mino_state_t *S)
 {
     mino_val_t *lhs  = regs[IMM_B];
     mino_val_t *kimm = IMM_KIMM;