+++ src/eval/bc/stencils/loop_int_lt.c
@@ -27,9 +27,9 @@
 
 #include "abi.h"
 
-mino_stencil_chain_t stencil_op_loop_int_lt(mino_val_t **regs,
-                                             mino_val_t **consts,
-                                             mino_state_t *S)
+void stencil_op_loop_int_lt(mino_val_t **regs,
+                             mino_val_t **consts,
+                             mino_state_t *S)
 {
     for (;;) {
         mino_val_t *vc = regs[IMM_A];
@@ -40,7 +40,7 @@ mino_stencil_chain_t stencil_op_loop_int_lt(mino_val_t **regs,
             /* Slow path: either operand non-int or NULL. */
             regs = mino_jit_loop_int_lt_slow(S, regs, IMM_A, IMM_B);
             if (regs == NULL) {
-                return (mino_stencil_chain_t){NULL, consts};
+                MINO_STENCIL_CHAIN_RETURN(NULL, consts, S);
             }
             if (((uintptr_t)regs & 1) != 0) {
                 regs = (mino_val_t **)((uintptr_t)regs & ~(uintptr_t)1);