+++ src/eval/bc/stencils/pop_env.c
@@ -10,9 +10,9 @@
 
 #include "abi.h"
 
-mino_stencil_chain_t stencil_op_pop_env(mino_val_t **regs,
-                                         mino_val_t **consts,
-                                         mino_state_t *S)
+void stencil_op_pop_env(mino_val_t **regs,
+                         mino_val_t **consts,
+                         mino_state_t *S)
 {
     regs = mino_jit_pop_env_slow(S, regs);
     MINO_STENCIL_CHAIN_RETURN(regs, consts, S);