+++ src/eval/bc/stencils/push_env.c
@@ -13,9 +13,9 @@
 
 #include "abi.h"
 
-mino_stencil_chain_t stencil_op_push_env(mino_val_t **regs,
-                                          mino_val_t **consts,
-                                          mino_state_t *S)
+void stencil_op_push_env(mino_val_t **regs,
+                          mino_val_t **consts,
+                          mino_state_t *S)
 {
     regs = mino_jit_push_env_slow(S, regs);
     MINO_STENCIL_CHAIN_RETURN(regs, consts, S);